chore: move js cache to file
